# output/openapi/openapi.json
{
    "openapi": "3.2.0",
    "info": {
        "title": "Fault-Tolerant University API",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": ""
        }
    ],
    "paths": {
...
        "/api/assignments/{id}/submit": {
            "post": {
                "operationId": "submitAssignment",
                "tags": [
                    "Submission"
                ],
                "responses": {
                    "201": {
                        "description": "Submission metadata"
                    },
                    "400": {
                        "description": "Invalid multipart or size"
                    },
                    "403": {
                        "description": "Enrollment required"
                    },
                    "409": {
                        "description": "Deadline passed or already submitted"
                    }
                },
                "summary": "Submit exactly one file, any format, at most 10,000,000 bytes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "integer"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "file"
                                ],
                                "additionalProperties": false,
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/api/submissions/{id}/download": {
            "get": {
                "operationId": "downloadSubmission",
                "tags": [
                    "Submission"
                ],
                "responses": {
                    "200": {
                        "description": "Binary document",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "summary": "Owner or administrator private download",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "integer"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ],
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "JWT": []
        }
    ],
...
}